pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-rl-trading-bot"
version = "1.0.0"
description = "Autonomous AI Trading Bot — UWE Digital Systems Project (UFCFXK-30-3)"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Mohamed Elkiky"}]

dependencies = [
    "numpy>=1.26,<2.0",
    "pandas>=2.1",
    "yfinance>=0.2.31",
    "gymnasium>=0.29",
    "torch>=2.1",
    "stable-baselines3>=2.2",
    "shimmy>=1.3",
    "scikit-learn>=1.3",
    "plotly>=5.18",
    "matplotlib>=3.8",
    "streamlit>=1.29",
    "pyyaml>=6.0",
    "rich>=13.7",
    "pytest>=7.4",
    "kaleido>=0.2.1",
    "ta>=0.11",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "certifi",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pre-commit"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
